Refactor AbstractFilterRepository to FilterRepositoryUtils
AbstractFilterRepository is refactored and renamed to FilterRepositoryUtils as a utility class. The README file has been updated to reflect this change, specifically mentioning the correct method of defining sort, limit, and offset in queries. The refactoring also involves changing the 'applyFilters' method to a static function that can be accessed without instantiating the class.